Another incremental checkin.
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1204 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/header/kernel/runtime/RexxBehaviour.hpp
@@ -50,7 +50,7 @@ void behaviour_setup (void);
 
  class RexxBehaviour : public RexxInternalObject {
   public:
-  void *operator new(size_t, short);
+  void *operator new(size_t, size_t);
   inline void *operator new(size_t size, void *ptr) {return ptr;};
   RexxBehaviour(size_t, PCPPM *);
   inline RexxBehaviour() {;};
